+++ io.openems.edge.victron/test/io/openems/edge/victron/battery/MyConfig.java
@@ -1,8 +1,7 @@
-package io.openems.edge.battery.victron;
+package io.openems.edge.victron.battery;
 
 import io.openems.common.test.AbstractComponentConfig;
 import io.openems.common.utils.ConfigUtils;
-import io.openems.edge.victron.battery.Config;
 
 @SuppressWarnings("all")
 public class MyConfig extends AbstractComponentConfig implements Config {
