+++ io.openems.edge.victron/test/io/openems/edge/victron/ess/MyConfig.java
@@ -1,9 +1,8 @@
-package io.openems.edge.batteryinverter.victron.ess.symmetric;
+package io.openems.edge.victron.ess;
 
 import io.openems.common.test.AbstractComponentConfig;
 import io.openems.common.utils.ConfigUtils;
 import io.openems.edge.common.type.Phase.SingleOrAllPhase;
-import io.openems.edge.victron.ess.Config;
 
 @SuppressWarnings("all")
 public class MyConfig extends AbstractComponentConfig implements Config {
