+++ io.openems.edge.victron/test/io/openems/edge/victron/meter/grid/MyConfig.java
@@ -1,9 +1,7 @@
-package io.openems.edge.pvinverter.victron;
+package io.openems.edge.victron.meter.grid;
 
 import io.openems.common.test.AbstractComponentConfig;
-import io.openems.common.types.MeterType;
 import io.openems.common.utils.ConfigUtils;
-import io.openems.edge.victron.pvinverter.Config;
 
 @SuppressWarnings("all")
 public class MyConfig extends AbstractComponentConfig implements Config {
@@ -14,7 +12,6 @@ public class MyConfig extends AbstractComponentConfig implements Config {
 		private boolean enabled;
 		private String modbusId;
 		private int modbusUnitId;
-		private MeterType type;
 
 		private Builder() {
 		}
@@ -41,11 +38,6 @@ public class MyConfig extends AbstractComponentConfig implements Config {
 
 		public Builder setModbusUnitId(int modbusUnitId) {
 			this.modbusUnitId = modbusUnitId;
-			return this;
-		}
-
-		public Builder setType(MeterType type) {
-			this.type = type;
 			return this;
 		}
 
@@ -98,11 +90,6 @@ public class MyConfig extends AbstractComponentConfig implements Config {
 	@Override
 	public int modbusUnitId() {
 		return this.builder.modbusUnitId;
-	}
-
-	@Override
-	public MeterType type() {
-		return this.builder.type;
 	}
 
 }
